# openapi.yaml
openapi: 3.1.0
info:
  title: GitHub Issues Gateway
  version: 1.0.0
  summary: A validated HTTP gateway over the GitHub Issues REST API.
  description: |
    A small service that fronts the GitHub Issues REST API for one configured
    repository (`GITHUB_OWNER/GITHUB_REPO`) and ingests that repository's
    webhooks.

    ## Authentication

    Clients of *this* API are unauthenticated -- it is intended to run on
    localhost or behind your own gateway. The service authenticates to GitHub
    on the caller's behalf using a fine-grained PAT supplied out-of-band in
    `GITHUB_TOKEN`; that is the `githubToken` security scheme below. It is
    declared on the routes that reach GitHub so the dependency is explicit in
    the contract, but it is **never** accepted from the client.

    `POST /webhook` is authenticated differently: GitHub signs each delivery
    with an HMAC-SHA256 of the raw body under `WEBHOOK_SECRET`, carried in
    `X-Hub-Signature-256`. That is the `webhookSignature` scheme.

    ## Deleting issues

    GitHub's REST API has no "delete issue" operation. Deletion is modelled as
    `PATCH /issues/{number}` with `{"state": "closed"}`.

    ## Pagination

    `GET /issues` and `GET /issues/{number}/comments` accept `page` and
    `per_page` (max 100) and return an RFC 8288 `Link` header whose `first`,
    `prev`, `next` and `last` URLs point back at **this** service, preserving
    the caller's filters.

    ## Rate limits

...

  contact:
    name: CMPE 272 HW2
...
servers:
...
paths:
...
  /webhook:
    post:
      tags: [webhooks]
      operationId: receiveWebhook
      summary: Receive a GitHub webhook delivery
      description: |
        Verifies `X-Hub-Signature-256` against `WEBHOOK_SECRET` with a
        constant-time comparison **before** parsing the body, persists the
        delivery, acks `204`, and interprets it in the background.

        Idempotent: `(X-GitHub-Delivery, event, action)` is the dedupe key, so
        GitHub's automatic retries and manual "Redeliver" both ack `204`
        without producing a second stored event.
      security:
        - webhookSignature: []
      parameters:
        - name: X-GitHub-Event
          in: header
          required: true
          description: Event name. One of `issues`, `issue_comment`, `ping`.
          schema: { type: string, enum: [issues, issue_comment, ping] }
        - name: X-GitHub-Delivery
          in: header
          required: false
          description: |
            GitHub's delivery GUID, used as the dedupe key. If absent, a SHA-256
            of the body is substituted so replays still dedupe.
          schema: { type: string, format: uuid }
        - name: X-Hub-Signature-256
          in: header
          required: true
          description: "`sha256=` followed by the hex HMAC-SHA256 of the raw body."
          schema: { type: string, pattern: "^sha256=[0-9a-fA-F]{64}$" }
      requestBody:
        required: true
        description: The raw GitHub event payload. Signed over byte-for-byte.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookPayload" }
            examples:
              issue_opened:
                summary: issues / opened
                value:
                  action: opened
                  issue:
                    number: 12
                    title: Rate limiter drops the Retry-After header
                    state: open
                  repository: { full_name: mikechau1/cmpe272-issues-gw }
                  sender: { login: mikechau1 }
              comment_created:
                summary: issue_comment / created
                value:
                  action: created
                  issue: { number: 12, title: Rate limiter drops the Retry-After header }
                  comment: { id: 3310028841, body: Confirmed on staging. }
                  repository: { full_name: mikechau1/cmpe272-issues-gw }
                  sender: { login: mikechau1 }
              ping:
                summary: ping
                value:
                  zen: Non-blocking is better than blocking.
                  hook_id: 552341190
                  hook: { type: Repository, events: [issues, issue_comment] }
      responses:
        "204":
          description: |
            Delivery accepted (or recognised as a duplicate and ignored).
            No body.
          headers:
            X-Request-ID: { $ref: "#/components/headers/XRequestId" }
        "400":
          description: Unknown event, unknown action, or malformed JSON.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                unsupported_event:
                  value:
                    error:
                      code: unsupported_event
                      message: "Unsupported webhook event 'push'. This gateway handles: issue_comment, issues, ping."
                      status: 400
                      request_id: 9bf31c7ff062936a96d3c8bd1f8f2ff3
                      details: { event: push, supported: [issue_comment, issues, ping] }
                unsupported_action:
                  value:
                    error:
                      code: unsupported_action
                      message: "Unsupported action 'exploded' for event 'issues'."
                      status: 400
                      request_id: c74d97b01eae257e44aa9d5bade97baf
        "401":
          description: The signature was missing, malformed, or did not match.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                bad_signature:
                  value:
                    error:
                      code: invalid_signature
                      message: "Invalid or missing X-Hub-Signature-256. The delivery body did not match an HMAC-SHA256 computed with WEBHOOK_SECRET."
                      status: 401
                      request_id: 70efdf2ec9b086079795c442636b55fb
        "413":
          description: The delivery body exceeded the configured size limit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /events:
    get:
      tags: [webhooks]
      operationId: listEvents
      summary: Recent webhook deliveries
      description: |
        The last N deliveries this service accepted, newest first. Intended for
        debugging and for asserting end-to-end webhook behaviour in tests.
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of deliveries to return.
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - name: event
          in: query
          required: false
          description: Restrict to one event type.
          schema: { type: string, enum: [issues, issue_comment, ping] }
      responses:
        "200":
          description: Recent deliveries, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StoredEvent" }
              examples:
                recent:
                  value:
                    - id: 42
                      delivery_id: 4f1a2b60-8c3d-11f0-9e21-9b4a2c1d7e55
                      event: issue_comment
                      action: created
                      issue_number: 12
                      repository: mikechau1/cmpe272-issues-gw
                      sender: mikechau1
                      status: processed
                      error: null
                      attempts: 1
                      timestamp: "2026-09-06T19:02:45.113Z"
                      processed_at: "2026-09-06T19:02:45.119Z"
                    - id: 41
                      delivery_id: 3ab0c14e-8c3d-11f0-8f77-1d2e3f4a5b6c
                      event: issues
                      action: opened
                      issue_number: 12
                      repository: mikechau1/cmpe272-issues-gw
                      sender: mikechau1
                      status: processed
                      error: null
                      attempts: 1
                      timestamp: "2026-09-06T18:20:12.004Z"
                      processed_at: "2026-09-06T18:20:12.010Z"
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  securitySchemes:
    githubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub fine-grained PAT
      description: |
        The server-side credential this gateway uses to call GitHub, supplied
        via the `GITHUB_TOKEN` environment variable. Declared so the dependency
        is visible in the contract; it is **never** read from the client
        request.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: |
        HMAC-SHA256 of the raw request body keyed with `WEBHOOK_SECRET`,
        formatted as `sha256=<hex>` and compared in constant time.
...
  headers:
...
    XRequestId:
      description: Correlation id, echoed from the request or generated.
      schema: { type: string, examples: [8f14e45fceea167a5a36dedd4bea2543] }
...
  schemas:
    Error:
      type: object
      description: The single error shape returned by every failing route.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, status]
          properties:
            code:
              type: string
              description: Stable, machine-readable identifier for the failure.
              examples: [validation_error, not_found, rate_limited, github_unauthorized]
            message:
              type: string
              description: Human-readable explanation, safe to show a developer.
            status:
              type: integer
              description: HTTP status of this response.
              minimum: 400
              maximum: 599
            request_id:
              type: string
              description: Correlation id; matches the `X-Request-ID` header and the logs.
            upstream_status:
              type: [integer, "null"]
              description: GitHub's status code, when the failure originated upstream.
            details:
              description: Structured specifics -- field errors, or GitHub's `errors` array.
              type: [array, object, "null"]
...
    WebhookPayload:
      type: object
      description: |
        GitHub's event payload. Only the fields this service reads are
        described; everything else is stored verbatim and ignored.
      properties:
        action:
          type: string
          description: Required for `issues` and `issue_comment`; absent on `ping`.
        issue:
          type: object
          properties:
            number: { type: integer }
            title: { type: string }
            state: { type: string }
        comment:
          type: object
          properties:
            id: { type: integer }
            body: { type: string }
        repository:
          type: object
          properties:
            full_name: { type: string, examples: [mikechau1/cmpe272-issues-gw] }
        sender:
          type: object
          properties:
            login: { type: string }
        zen: { type: string, description: Present on `ping` only. }
        hook_id: { type: integer, description: Present on `ping` only. }
    StoredEvent:
      type: object
      description: One accepted webhook delivery as recorded by this service.
      required: [id, event, timestamp]
      properties:
        id: { type: integer, description: Monotonic local id. }
        delivery_id: { type: string, description: GitHub's `X-GitHub-Delivery` GUID. }
        event: { type: string, enum: [issues, issue_comment, ping] }
        action: { type: [string, "null"], examples: [opened, closed, created] }
        issue_number: { type: [integer, "null"] }
        repository: { type: [string, "null"] }
        sender: { type: [string, "null"] }
        status:
          type: string
          enum: [received, processed, failed]
          description: "`failed` marks a poison delivery, parked with its error."
        error: { type: [string, "null"] }
        attempts: { type: integer }
        timestamp: { type: string, format: date-time, description: When the delivery was accepted. }
        processed_at: { type: [string, "null"], format: date-time }
...
  responses:
    BadRequest:
      description: The request failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            invalid_state:
              value:
                error:
                  code: validation_error
                  message: "Request validation failed -- query.state: Input should be 'open', 'closed' or 'all'"
                  status: 400
                  request_id: 1679091c5a880faf6fb5e6087eb1b2dc
                  details:
                    - field: query.state
                      message: "Input should be 'open', 'closed' or 'all'"
                      type: literal_error
                      input: sideways
            per_page_too_large:
              value:
                error:
                  code: validation_error
                  message: "Request validation failed -- query.per_page: Input should be less than or equal to 100"
                  status: 400
                  request_id: 8f14e45fceea167a5a36dedd4bea2543
...
    NotConfigured:
      description: A required environment variable is unset.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            missing_repo:
              value:
                error:
                  code: not_configured
                  message: "GITHUB_OWNER and GITHUB_REPO must both be set for the gateway to know which repository to operate on."
                  status: 503
                  request_id: c74d97b01eae257e44aa9d5bade97baf
            missing_secret:
              value:
                error:
                  code: not_configured
                  message: "WEBHOOK_SECRET is not set, so delivery signatures cannot be verified. The gateway refuses unsigned webhook traffic."
                  status: 503
                  request_id: 70efdf2ec9b086079795c442636b55fb